# demoAPI/demoAPI/openapi.yaml
openapi: 3.0.1
info:
  title: demoAPI
  version: v1
  host: carbide-ward-303004.appspot.com
paths:
  /Location:
    get:
      tags:
        - Location
      responses:
        '200':
          description: Success
    post:
      tags:
        - Location
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
          text/json:
            schema:
              $ref: '#/components/schemas/Location'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '200':
          description: Success
  '/Location/{id}':
    get:
      tags:
        - Location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
...
components:
  schemas:
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int32
        location_name:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
...